pyproject.toml
[project]
name = "voice-robot"
version = "1.0.0"
description = "Voice Robot — Tauri v2 + Vue 3 + FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "pydantic>=2.0",
    "websockets<14",
    "dashscope>=1.25.15",
    "livekit-api>=1.1.0",
]

[tool.uv]
dev-dependencies = []

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
